Remove console.log and replace with mlog. Rename file
